Update load behaviour to only derive public key if both keys are present.
This commit fixes the corner case where a consuming service, for
example, sets the c2 public key but not the device secret key, then
syncs to storage, then reloads from storage. Previously a shared key
would be derived under these conditions. Since all zeros will be written
out the loaded key material will also be all zero so we check for this
to ensure the key is initialized.