Implement code size count from makefile.
This commit uses the portable POSIX solution for counting the size of
the output library (.a), namely wc -c.
The output is specifically produced in the successful "end of make"
message.
Alternatively, one can now explicitly "make codesize".

We should specifically note that: code size will depend on target
architecture (x86_64 has 8-byte pointers, for example) and does not
account for the extraction of the various elf-objects. That is, this
code currently gives an upper bound estimate of the size INCLUDING **ELF**
metadata and symbols for each and every file in the archive. A release
image would strip all this information.